+++ examples/data-sources/nxos_default_qos_policy_interface_in_policy_map/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_default_qos_policy_interface_in_policy_map" "example" {
-  interface_id = "eth1/10"
-}