+++ examples/resources/nxos_default_qos_policy_interface_in/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_default_qos_policy_interface_in.example
-  identity = {
-    "interface_id" : "<interface_id>"
-  }
-}