+++ examples/resources/nxos_default_qos_policy_interface_in/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_default_qos_policy_interface_in.example
-  id = "<interface_id>"
-}