+++ examples/resources/nxos_default_qos_policy_interface_in/resource.tf
@@ -1,3 +1,0 @@
-resource "nxos_default_qos_policy_interface_in" "example" {
-  interface_id = "eth1/10"
-}