+++ examples/resources/nxos_default_qos_policy_interface_in_policy_map/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_default_qos_policy_interface_in_policy_map.example "<interface_id>"